+++ other-tests/other/other_test.cljs
@@ -1,0 +1,6 @@
+(ns other.other-test
+  (:require [cljs.test :as t]))
+
+(t/deftest other-should-run
+  (t/testing "this should run in the other dir"
+    (t/is (= 1 1))))