+++ test/example/yes_test.cljs
@@ -4,3 +4,7 @@
 (t/deftest should-run
   (t/testing "this should run"
     (t/is (= 1 1))))
+
+(t/deftest ^:integration maybe-run
+  (t/testing "this may run"
+    (t/is (= 2 2))))